05-django/solutions/exercise-5-filters/requirements.txt
django
django-filter
djangorestframework
markdown
